refactor: fix package json after rebase
